moved sysid logging to subfolder
