+++ src/main/java/frc/robot/commands/Characterize.java
@@ -3,7 +3,7 @@ package frc.robot.commands;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.Drivetrain;
-import frc.robot.SysIdDrivetrainLogger;
+import frc.robot.sysid.SysIdDrivetrainLogger;
 
 public class Characterize extends CommandBase {
     private final Drivetrain m_drivetrain;
