+++ src/main/java/frc/robot/sysid/SysIdLogger.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.sysid;
 
 import java.util.ArrayList;
